+++ gladier_xpcs/flows/__init__.py
@@ -1,11 +1,11 @@
-from .flow_online import XPCSOnlineFlow
-from .flow_gpu import XPCSGPUFlow
+from .flow_eigen import XPCSEigen
+from .flow_boost import XPCSBoost
 from .flow_reprocess import XPCSReprocessingFlow
 
 
 __all__ = [
-    'XPCSOnlineFlow',
-    'XPCSGPUFlow',
+    'XPCSEigen',
+    'XPCSBoost',
     'XPCSReprocessingFlow'
     ]
