+++ gladier_xpcs/tools/__init__.py
@@ -1,7 +1,7 @@
 from .transfer_from_clutch_to_theta import TransferFromClutchToTheta
 from .pre_publish import PrePublish
 
-from .corr import EigenCorr
+from .eigen_corr import EigenCorr
 from .xpcs_boost_corr import BoostCorr
 
 from .plot import MakeCorrPlots
