+++ scripts/xpcs_online_boost_client.py
@@ -9,7 +9,7 @@ import argparse
 import os
 import pathlib
 
-from gladier_xpcs.flows import XPCSGPUFlow
+from gladier_xpcs.flows import XPCSBoost
 from gladier_xpcs.deployments import deployment_map
 
 
@@ -113,7 +113,7 @@ if __name__ == '__main__':
         }
     }
 
-    corr_flow = XPCSGPUFlow()
+    corr_flow = XPCSBoost()
     corr_run_label = pathlib.Path(hdf_name).name[:62]
     flow_run = corr_flow.run_flow(flow_input=flow_input, label=corr_run_label)
